+++ physicsnemo/nn/__init__.py
@@ -92,6 +92,7 @@ from .module.embedding_layers import (
     PositionalEmbedding,
     SinusoidalTimestepEmbedding,
 )
+from .module.flare_attention import FLARE
 from .module.fourier_layers import (
     FourierFilter,
     FourierLayer,
@@ -107,6 +108,13 @@ from .module.fully_connected_layers import (
     ConvNdKernel1Layer,
     FCLayer,
     Linear,
+)
+from .module.gale import (
+    GALE,
+    GALE_FA,
+    GALEBlock,
+    GALEStructuredMesh2D,
+    GALEStructuredMesh3D,
 )
 from .module.group_norm import GroupNorm, get_group_norm
 from .module.gumbel_softmax import GumbelSoftmax, gumbel_softmax